Move BinarySerializer to the correct namespace
