+++ SAGESharp.Tests/IO/Binary/BinarySerializerTests.cs
@@ -13,7 +13,7 @@ using System;
 using System.IO;
 using System.Reflection;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Binary
 {
     class BinarySerializerTests
     {
